Fix flag() with rest() support with no arg() (#4)
Rest was triggered when all defined arguments were set, but in a
scenario where there are no defined arguments and only a `rest()` to
vacuum up arguments flags were never processed even if they were first.

The Parser now attempts to parse first and if the command falls through
to interpreting the result as an argument, then the check for the number
of defined arguments found is run. This means the parser needs to be
walked back one place as it has already processed the first argument of
the rest at that point.

It is assumed that the first non-defined argument found is the start of
the rest arguments and all remaining arguments belong to the rest. This
means even valid flags after the first rest argument are considered rest
arguments, not flags.